Added constants, fixed date class

+++ datetime.h
@@ -7,7 +7,7 @@ public:
 	uint8_t dayOfWeek;
 	uint8_t day;
 	uint8_t month;
-	uint8_t year;
+	uint16_t year;
 	
 };
